feat: Add quick-add expense icon to dashboard category rows
Added a reusable expense form dialog that can be triggered from anywhere in the app, including a convenient PlusCircle icon on each category row in the dashboard's budget summary table.

## Changes

### New Component
- **expense-form-dialog.tsx**: Fully reusable expense form dialog
  - Pre-populate category via `preSelectedCategoryId` prop
  - Supports current fund filter context
  - Complete validation and error handling
  - Success callback for data refresh

### Modified Components
- **dashboard-view.tsx**:
  - Added PlusCircle icon next to each category name
  - Icon appears on row hover with smooth opacity transition
  - Integrated ExpenseFormDialog with pre-selected category
  - Added refresh trigger to update data after expense creation

- **expenses-view.tsx**:
  - Refactored to use new ExpenseFormDialog component
  - Removed 250+ lines of duplicate form code
  - Maintained all existing edit/delete functionality

### Documentation
- **CLAUDE.md**: Added "Quick-Add Expense Pattern" section
- **feature-quick-add-expense-from-dashboard.md**: Complete implementation plan and details

## Features
- ✅ Hover-activated icon (subtle, non-intrusive)
- ✅ Pre-selected & disabled category field
- ✅ Fund relationship validation
- ✅ Auto data refresh after submission
- ✅ Success toast notifications
- ✅ Clean state management

## UX Enhancements
- Icon only visible on row hover
- Smooth opacity transition
- Tooltip for clarity
- Helper text for pre-selected category
- Automatic cleanup on dialog close
